fix(assets): change interest_rate type from int to float in Asset and AssetDto classes for improved precision

+++ backend/app/contexts/assets/domain/entities/asset.py
@@ -6,4 +6,4 @@ from app.contexts.shared.domain.entity import Entity
 
 
 class Asset(Entity):
-    interest_rate: Annotated[int, Field(description="The interest rate of the asset")]
+    interest_rate: Annotated[float, Field(description="The interest rate of the asset")]
